+++ single_include/nlohmann/json.hpp
@@ -10239,7 +10239,7 @@ class binary_reader
                             {
                                 return false;
                             }
-                            result = number;
+                            result = static_cast<std::size_t>(number);
                             return true;
                         }
                         case '[':
